Added OpenAL library and dependencies

+++ apps/my_app.h
@@ -30,7 +30,7 @@ class MyApp : public cinder::app::App {
   bool CheckTopBottomCollision(BrickBreaker::ball ball, BrickBreaker::brick brick);
   bool CheckSideCollision(BrickBreaker::ball ball, BrickBreaker::brick brick);
   void DrawMenuScreen();
-  void SelectLevel(int level_number);
+  void SelectLevel(size_t level_number);
   void GenerateLevels();
   int GetLevelClicked(ci::vec2 loc_clicked);
   template<typename C> void PrintText(const std::string &text, const C &color, const cinder::ivec2 &size,
@@ -45,7 +45,7 @@ class MyApp : public cinder::app::App {
   int last_click_count_;
   int menu_grid_width_;
   int menu_grid_height_;
-  int current_level_;
+  size_t current_level_;
 
   uint64_t time_;
   uint64_t last_collision_time_;
